requirements.txt
asyncio
aiofiles
pydantic
numpy
cryptography
python-dotenv
pyTelegramBotAPI
requests
psutil
sentence-transformers
torch
llama-cpp-python